+++ desk/src/components/view-controls/filter/index.ts
@@ -1,3 +1,3 @@
-export { default as FilterDrilldown } from "./FilterDrilldown.vue";
+export { default as Filter } from "./Filter.vue";
 export { normalizeFilters } from "./filterCore";
 export type { ActiveFilter, FilterCondition, FilterField } from "./filterCore";
